<!--  /**   * @FileName	: MeetingController.java   * @Date		: 2019. 11. 27.   * @Author		: 장윤석   * @프로그램 설명 : free_board_vo   */  -->

+++ branches/teamB_project/src/main/java/com/ync/project/domain/Free_BoardVO.java
@@ -1,6 +1,8 @@
 package com.ync.project.domain;
 
+
 import java.util.Date;
+
 import lombok.Data;
 
  /**
@@ -21,8 +23,8 @@ import lombok.Data;
 @Data
 public class Free_BoardVO {
 	
-	private int meeting_num;		//모임 번호
-	private int free_board_num;		//게시글 번호
+	private Long meeting_num;		//모임 번호
+	private Long free_board_num;		//게시글 번호
 	
 	private String title;			//게시글 제목
 	private String category;		//글 분류
@@ -30,7 +32,7 @@ public class Free_BoardVO {
 	private String files;			//파일
 	
 	private Date write_Date;		//게시글 작성일
-	private int views;				//조회수
+	private Long views;				//조회수
 	private String userid;			//작성자 아이디
 	
 }
